refactor(lint): clippy sweep — zero warnings on --all-targets
Pre-release cleanup for 0.8.0. `cargo clippy --all-targets` (default
features) now reports zero warnings. Roughly half landed via
`cargo clippy --fix`; the rest hand-fixed.

Hand-fixes of note:
- probability.rs: `impl Into` → `impl From` (from_over_into); From gives
  Into for free, so the conversion is source-compatible.
- session/mod.rs: `is_open` takes `&str` not `&Cow` (ptr_arg); the two
  `if cache_on { Vec::new() } else { Vec::new() }` collapse to `Vec::new()`.
- candidates.rs: manual `.min(MAX).max(MIN)` → `.clamp` (input is a
  non-negative finite `abs`, so the NaN-handling difference can't arise).
- doc-lint reflows (doc_lazy_continuation / overindented list items) in
  candidates.rs, repetition.rs, session/mod.rs, README.md — the session
  one was a line-leading `+` being parsed as a markdown bullet.
- probe_hook.rs: a `SnapshotLog` type alias for the twice-used capture type.

`#[allow]` with a site reason where the lint wanted an API break or a
semantics-losing rewrite:
- candidates.rs `as_ref`/`as_mut` (should_implement_trait): renaming is a
  public break, and `AsMut` is the wrong contract (side-effecting reset).
  Noted for 1.0 in one_dot_oh_wishlist.md.
- session/mod.rs three private helpers + a test fixture (type_complexity):
  one-off tuples whose positional meaning a type alias would hide.
- repetition.rs test (same_item_push): the pushes interleave with sampling.

+++ src/log.rs
@@ -52,7 +52,7 @@ pub(crate) fn log_level_from_raw(level: ggml_log_level) -> LogLevel {
         ggml_log_level_GGML_LOG_LEVEL_WARN => LogLevel::Warn,
         ggml_log_level_GGML_LOG_LEVEL_ERROR => LogLevel::Error,
         ggml_log_level_GGML_LOG_LEVEL_CONT => LogLevel::Cont,
-        other => LogLevel::Other(other as u32),
+        other => LogLevel::Other(other),
     }
 }
